Build into the public directory on GitLab
The "static" source files used to live in the `public/` directory. The
build used to build into `dist/`. GitLab Pages looks in the `public/`
directory for a built static site to serve. So we free up the
`public/` directory by renaming it to `app/`. Then we alter the build
process on GitLab to build into the `public/` directory.

[From the GitLab docs][1]:

> "Any static content which will be served by GitLab Pages must be
> placed under a `public/` directory"

[1]: https://docs.gitlab.com/ce/user/project/pages/introduction.html